chore: organize list plugin like other plugins
