+++ packages/runtime/src/components/builtin/Text/EditableText/editable-text.tsx
@@ -22,10 +22,10 @@ import { BlockType, RichTextDAO, richTextDTOtoDAO } from '../../../../controls'
 import { Leaf } from '../components/Leaf'
 import { Element } from '../components/Element'
 import { useSyncWithBuilder } from './useSyncWithBuilder'
-import { onKeyDown, withList } from '../../../../slate/ListPlugin'
 import isHotkey from 'is-hotkey'
 import { useBuilderEditMode } from '../../../../runtimes/react'
 import { BuilderEditMode } from '../../../../state/modules/builder-edit-mode'
+import { onKeyDown, withList } from '../../../../slate'
 
 type Props = {
   id?: ElementIDValue
