--- packages/runtime/src/slate/ListPlugin/lib/index.ts
@@ -1,5 +1,0 @@
-export * from './wrapList'
-export * from './unwrapList'
-export * from './toggleList'
-export * from './dedent'
-export * from './indent'
