--- packages/runtime/src/slate/ListPlugin/test/types.ts
@@ -1,21 +1,0 @@
-export const Editor = 'editor'
-export const Fragment = 'fragment'
-export const Anchor = 'anchor'
-export const Focus = 'focus'
-export const Cursor = 'cursor'
-export const Selection = 'selection'
-
-export const Text = 'text'
-
-export const Paragraph = 'paragraph'
-export const Heading1 = 'heading1'
-export const Heading2 = 'heading2'
-export const Heading3 = 'heading3'
-export const Heading4 = 'heading4'
-export const Heading5 = 'heading5'
-export const Heading6 = 'heading6'
-export const BlockQuote = 'blockquote'
-export const Ordered = 'ordered'
-export const Unordered = 'unordered'
-export const ListItem = 'listitem'
-export const ListItemChild = 'listitemchild'
