+++ packages/runtime/src/slate/ListPlugin/toggleList.ts
@@ -1,7 +1,7 @@
 import { Editor, Node, Path, Transforms } from 'slate'
-import { BlockType } from '../../../controls'
+import { BlockType } from '../../controls'
+import { ElementUtils } from '../utils/element'
 import { unwrapList } from './unwrapList'
-import { ElementUtils } from './utils/element'
 import { LocationUtils } from './utils/location'
 import { wrapList } from './wrapList'
