+++ packages/runtime/src/slate/ListPlugin/unwrapList.ts
@@ -1,7 +1,7 @@
 import { Editor, Transforms, Path, Node } from 'slate'
-import { BlockType } from '../../../controls'
-import { LIST_ITEM_CHILD_POSITION, LIST_ITEM_LIST_POSITION } from '../constants'
-import { EditorUtils } from './utils/editor'
+import { BlockType } from '../../controls'
+import { EditorUtils } from '../utils/editor'
+import { LIST_ITEM_CHILD_POSITION, LIST_ITEM_LIST_POSITION } from './constants'
 import { filterForSubtreeRoots } from './utils/filterForSubtreeRoots'
 import { getSelectedListItems } from './utils/getSelectedListItems'
