+++ packages/runtime/src/slate/index.ts
@@ -1,2 +1,3 @@
 export * from './ListPlugin'
 export * from './TypographyPlugin'
+export * from './BlockPlugin'
